Move server directory logic to Server & BukkitServer
